Update to functions
Only change to umls_semantic_occurrences function is that now it accepts an array of semantic types to filter by. This works now, because I have updated the filter_mesh_by_concept function to take in an array of semantic types to filter on. I change the query that we are sending to mysql and get the propoer one by iterating through the array of semantic types then doing some string manipulation. Lastly, I have included the fucntion occurences_to_itemsets, which gets the disease_occurances sparse matrix into a dataframe ready to be plugged into the apriori function. Let me know what you think.